Create ClientSession and first API request mock
The ClientSession class wraps a session object of the requests library.
It will be used to make HTTP requests to the mock REST API.

The API is mocked using requests-mock library. A session adapter is
created and mock sever responses are registered within it. This adapter
is then mounted on the session object in ClientSession. All HTTP
requests are then sent through the adapter, which replies with the
registered mock responses.

This commit implements the initial skeleton of this system, which will
be extended by upcoming changes.