Implement CreateTask command
Added new command for creating Tasks. Tasks are represented by Task
objects and are held by ServerMock for mocking purposes. ServerMock will
actually create a new Task and return its ID.

ServerMock now also stores a list of robot names. Requesting a Task for
unknown robot will cause the request to fail. At this point there is no
way to query the robot list, but a command for this will be added in
upcoming commits.