Implement a custom help command
The built-in "help" command offered by thee cmd library doesn't seem to
work for decorated functions. This resulted in most commands losing
their auto-generated help messages, which were supposed to be composed
using their docstrings.

To solve this problem, a custom "help" command was implemented, which
uses a command dict to access the command object and ask it for its
description, stored in its argparser. This also resulted in refactoring
of command creation and storage in UserInterface class.