Replace request URL magic strings with function calls
So far, both ServerMock and ClientSession used magic strings to define
request URLs. This was prone to errors, as one typo could break the
request's processing.

To address that issue, a new file "api.py" was created. It contains
classes that define URL components and can produce request URLs for
supported requests. Using the same URL generation function in both
request definition and the mock API eliminates the danger of URL
mismatch between the two.